+++ docs/trait.impl/core/ops/bit/trait.Not.js
@@ -1,9 +1,9 @@
 (function() {
-    const implementors = Object.fromEntries([["arrow_schema",[["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/1.95.0/core/ops/bit/trait.Not.html\" title=\"trait core::ops::bit::Not\">Not</a> for <a class=\"struct\" href=\"arrow_schema/struct.SortOptions.html\" title=\"struct arrow_schema::SortOptions\">SortOptions</a>",0]]],["hsc",[["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/1.95.0/core/ops/bit/trait.Not.html\" title=\"trait core::ops::bit::Not\">Not</a> for <a class=\"enum\" href=\"hsc/snapshots/enum.Stats2Save.html\" title=\"enum hsc::snapshots::Stats2Save\">Stats2Save</a>",0]]]]);
+    const implementors = Object.fromEntries([["hsc",[["impl <a class=\"trait\" href=\"https://doc.rust-lang.org/1.96.0/core/ops/bit/trait.Not.html\" title=\"trait core::ops::bit::Not\">Not</a> for <a class=\"enum\" href=\"hsc/snapshots/enum.Stats2Save.html\" title=\"enum hsc::snapshots::Stats2Save\">Stats2Save</a>",0]]]]);
     if (window.register_implementors) {
         window.register_implementors(implementors);
     } else {
         window.pending_implementors = implementors;
     }
 })()
-//{"start":59,"fragment_lengths":[291,277]}
+//{"start":59,"fragment_lengths":[276]}